Add CriteriaJudge for evaluating Purple agent performance (Phase 3.7)
Implement the CriteriaJudge class which evaluates Purple agent performance
against scenario-defined evaluation criteria. Supports both programmatic
evaluators (Python functions) and LLM-based evaluation with structured output.

New modules:
- src/green/judge.py: CriteriaJudge class with parallel evaluation, score
  scaling, error handling, and TaskUpdateEmitter integration
- src/green/judge_models.py: LLMEvaluationResult Pydantic model
- src/green/prompts/evaluation_prompts.py: Prompt templates and context
  builders for LLM-based criterion evaluation

Key features:
- Parallel criterion evaluation via asyncio.gather()
- Score scaling to match criterion max_score
- Graceful error handling with 0-score fallback
- TaskUpdateEmitter integration for observability
- Warning on missing evaluator references

Breaking change:
- Changed score types from int to float in results.py and updates.py to
  support accurate score scaling

Tests: 64 new tests (832 total passing)

AI generated commit message

+++ src/common/agentbeats/updates.py
@@ -301,8 +301,8 @@ class CriterionEvaluatedUpdate(BaseModel):
     criterion_id: str = Field(..., description="Criterion identifier")
     criterion_name: str = Field(..., description="Human-readable criterion name")
     dimension: str = Field(..., description="Scoring dimension")
-    score: int = Field(..., ge=0, description="Achieved score")
-    max_score: int = Field(..., ge=0, description="Maximum possible score")
+    score: float = Field(..., ge=0, description="Achieved score")
+    max_score: float = Field(..., ge=0, description="Maximum possible score")
     evaluation_method: Literal["programmatic", "llm"] = Field(
         ..., description="Evaluation method used"
     )
@@ -343,8 +343,8 @@ class AssessmentCompletedUpdate(BaseModel):
     total_turns: int = Field(..., ge=0, description="Total turns taken")
     total_actions: int = Field(..., ge=0, description="Total actions taken")
     duration_seconds: float = Field(..., ge=0, description="Duration in seconds")
-    overall_score: int = Field(..., ge=0, description="Final overall score")
-    max_score: int = Field(..., ge=0, description="Maximum possible score")
+    overall_score: float = Field(..., ge=0, description="Final overall score")
+    max_score: float = Field(..., ge=0, description="Maximum possible score")
 
 
 class ErrorOccurredUpdate(BaseModel):
@@ -764,8 +764,8 @@ class TaskUpdateEmitter:
         criterion_id: str,
         criterion_name: str,
         dimension: str,
-        score: int,
-        max_score: int,
+        score: float,
+        max_score: float,
         evaluation_method: Literal["programmatic", "llm"],
     ) -> TaskStatusUpdateEvent:
         """Emit a criterion evaluated update.
@@ -797,8 +797,8 @@ class TaskUpdateEmitter:
         total_turns: int,
         total_actions: int,
         duration_seconds: float,
-        overall_score: int,
-        max_score: int,
+        overall_score: float,
+        max_score: float,
     ) -> TaskStatusUpdateEvent:
         """Emit an assessment completed update.
